+++ dds-agent/src/Options.h
@@ -45,7 +45,7 @@ namespace dds
             }
 
             ECommands m_Command{ cmd_start };
-            size_t m_slots{ 0 };
+            uint32_t m_slots{ 0 };
             std::string m_groupName;
         } SOptions_t;
 
@@ -69,7 +69,7 @@ namespace dds
                                   "   start: \tStart dds-agent\n"
                                   "   clean: \tCleaning");
             options.add_options()(
-                "slots,s", bpo::value<size_t>(&_options->m_slots), "Defines a number of task slots per agent.");
+                "slots,s", bpo::value<uint32_t>(&_options->m_slots), "Defines a number of task slots per agent.");
             options.add_options()(
                 "group-name,g", bpo::value<std::string>(&_options->m_groupName), "Defines a group name of the agent.");
